+++ apps/web/src/components/settings/BAppSettings.vue
@@ -14,11 +14,13 @@ import { storeToRefs } from "pinia";
 import { computed, onMounted } from "vue";
 import { useI18n } from "vue-i18n";
 import BOrigins from "@/components/settings/BOrigins.vue";
+import BTimeZone from "@/components/settings/BTimeZone.vue";
 import { useSettingsStore } from "@/stores/settings.ts";
 
 const { t, availableLocales } = useI18n();
 const store = useSettingsStore();
-const { loading, error, language, interval } = storeToRefs(store);
+const { loading, error, language, interval, timezone, prefix } =
+	storeToRefs(store);
 const message = useMessage();
 
 const locales = computed(() =>
@@ -43,7 +45,7 @@ async function handleUpdateAliases() {
 }
 
 onMounted(async () => {
-	await store.getOrigins();
+	await Promise.all([store.getPrefix(), store.getOrigins()]);
 });
 </script>
 
@@ -63,6 +65,10 @@ onMounted(async () => {
       </NFormItem>
     </div>
 
+    <div class="w-full max-w-sm">
+      <BTimeZone v-model="timezone" :default="timezone" />
+    </div>
+
     <div class="flex gap-1">
       <div class="w-full flex gap-1 max-w-sm">
         <NFormItem :label="$t('interval')" class="w-full">
@@ -70,7 +76,11 @@ onMounted(async () => {
               v-model:value="interval"
               :show-button="false"
               class="w-full"
-          />
+          >
+            <template #prefix>
+              ms
+            </template>
+          </NInputNumber>
         </NFormItem>
       </div>
 
@@ -93,7 +103,9 @@ onMounted(async () => {
     </div>
 
     <BOrigins
-        v-model="store.settings.origins"
+        v-model:origins="store.settings.origins"
+        v-model:prefix="prefix"
+        @update:prefix="store.setPrefix"
         @handleChange="store.setOrigin"
     />
 
@@ -121,6 +133,7 @@ onMounted(async () => {
             type="primary"
             @click="handleUpdateAliases"
             :loading="loading"
+            :disabled="loading"
             class="min-w-[120px]"
         >
           {{ $t('set') }}